Squelch a signed vs. unsigned comparison warning on FreeBSD 10.
FD_SETSIZE might be defined as an unsigned, but sock_info->sock will be
an int on UN*Xes.  It's not going to be negative (that's used for errors
from calls that return FDs, and if we got an error we wouldn't have an
FD on which to select), so just cast both sides to unsigned int.

+++ rpcapd/rpcapd.c
@@ -893,7 +893,8 @@ accept_connections(void)
 	{
 		if (sock_info->sock + 1 > num_sock_fds)
 		{
-			if (sock_info->sock + 1 > FD_SETSIZE)
+			if ((unsigned int)(sock_info->sock + 1) >
+			    (unsigned int)FD_SETSIZE)
 			{
 				rpcapd_log(LOGPRIO_ERROR, "Socket FD is too bit for an fd_set");
 				exit(2);
